refactor: moves logic to the hook and awaits promises
